+++ src/test/scala/millfork/test/emu/EmuZ80Run.scala
@@ -24,6 +24,8 @@ import org.scalatest.Matchers
 class EmuZ80Run(cpu: millfork.Cpu.Value, nodeOptimizations: List[NodeOptimization], assemblyOptimizations: List[AssemblyOptimization[ZLine]]) extends Matchers {
   def inline: Boolean = false
 
+  def optimizeForSize: Boolean = false
+
   private val TooManyCycles: Long = 1500000
 
   def apply(source: String): MemoryBank = {
@@ -38,6 +40,7 @@ class EmuZ80Run(cpu: millfork.Cpu.Value, nodeOptimizations: List[NodeOptimizatio
     val platform = EmuPlatform.get(cpu)
     val extraFlags = Map(
       CompilationFlag.InlineFunctions -> this.inline,
+      CompilationFlag.OptimizeForSize -> this.optimizeForSize,
       CompilationFlag.EmitIllegals -> (cpu == millfork.Cpu.Z80),
       CompilationFlag.LenientTextEncoding -> true)
     val options = CompilationOptions(platform, millfork.Cpu.defaultFlags(cpu).map(_ -> true).toMap ++ extraFlags, None, 0, JobContext(log, new LabelGenerator))